feat: refactor all routing logic to updates.js
