+++ components/UpdatesWithData.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { graphql } from '@apollo/client/react/hoc';
-import { cloneDeep, omitBy } from 'lodash';
+import { cloneDeep } from 'lodash';
 import { FormattedMessage } from 'react-intl';
 
 import { API_V2_CONTEXT, gqlV2 } from '../lib/graphql/helpers';
@@ -15,8 +15,6 @@ import StyledButton from './StyledButton';
 import { H1, P } from './Text';
 import Updates from './Updates';
 
-const ROUTE_PARAMS = ['collectiveSlug', 'offset'];
-
 class UpdatesWithData extends React.Component {
   static propTypes = {
     collective: PropTypes.object,
@@ -26,7 +24,8 @@ class UpdatesWithData extends React.Component {
     LoggedInUser: PropTypes.object,
     data: PropTypes.object,
     fetchMore: PropTypes.func,
-    router: PropTypes.object,
+    onChange: PropTypes.func,
+    query: PropTypes.object,
   };
 
   constructor(props) {
@@ -45,15 +44,8 @@ class UpdatesWithData extends React.Component {
     }
   }
 
-  updateQuery = (router, newParams) => {
-    const query = omitBy({ ...router.query, ...newParams }, (value, key) => !value || ROUTE_PARAMS.includes(key));
-    const pathname = router.asPath.split('?')[0];
-    return router.push({ pathname, query });
-  };
-
   render() {
-    const { data, LoggedInUser, collective, compact, router } = this.props;
-    const query = router.query;
+    const { data, LoggedInUser, collective, compact, onChange, query } = this.props;
 
     if (data.error) {
       return <Error message={data.error.message} />;
@@ -84,15 +76,7 @@ class UpdatesWithData extends React.Component {
             )}
           </Flex>
         )}
-        <UpdateFilters
-          values={query}
-          onChange={queryParams =>
-            this.updateQuery(router, {
-              ...queryParams,
-              offset: null,
-            })
-          }
-        />
+        <UpdateFilters values={query} onChange={onChange} />
         <Box mt={4} mb={5}>
           <Updates
             collective={collective}
@@ -151,8 +135,8 @@ const getUpdatesVariables = props => {
     offset: 0,
     limit: props.limit || UPDATES_PER_PAGE * 2,
     includeHostedCollectives: props.includeHostedCollectives || false,
-    orderBy: { field: 'CREATED_AT', direction: props.router.query?.orderBy === 'oldest' ? 'ASC' : 'DESC' },
-    searchTerm: props.router.query?.searchTerm,
+    orderBy: { field: 'CREATED_AT', direction: props.query?.orderBy === 'oldest' ? 'ASC' : 'DESC' },
+    searchTerm: props.query?.searchTerm,
   };
 };
